[Impeller] Statically check stage interfaces for precision mismatches. (flutter/engine#54375)
Fixes https://github.com/flutter/flutter/issues/147078

+++ engine/src/flutter/impeller/entity/shaders/tiled_texture_fill.frag
@@ -18,7 +18,7 @@ uniform FragInfo {
 }
 frag_info;
 
-in highp vec2 v_texture_coords;
+in mediump vec2 v_texture_coords;
 
 out f16vec4 frag_color;
